Fix crash in ncclCommDestroy.
Since we now use comm->myParams->stream to store the NCCL stream, we
need to make sure that comm->params is not freed by intraRank 0
before the call to ncclCommDestroy. Instead of intraRank 0, the
intra-process shared resources are now freed last by the last rank to
call ncclCommDestroy, and after cudaStreamDestroy.