Remove per-datatype versions of Bcast and AllGather.
Bcast and Allgather only copy data, so they should not be dependent
on the datatype. Keeping only the int8_t version saves binary size,
and prepares for the inclusion of Scatter/Gather/Alltoall/P2P which
should also only have a single version and not cause too much size
increase.