Enable the user to reuse the same <ip>:<port> as the comm id of a next run
The listening socket at the root which is bound to the comm id can be in TIME_WAIT mode even after closing. And it can take some time for it to be really closed. This could lead to an "Address already in use" error in the next run if the same comm id is used.