Fix a bug on Skylake systems related to PCIe path
Newer versions of NVML returns device bus ID with upper case letters,
while /sys/bus/pci/device/<bus_id> paths use lower case.

Previous systems (like Broadwell) happen to be immune to this problem because the
switch IDs are formed of numbers, while in the case of Skylake the
switch IDs fall in the range of letters (a to f).