[bootstrap] keep only a handful of sockets open at any given time
This commit borrows from David Addison's simple bootstrap AllGather ring
patch and goes beyond that to ensure that a fixed number of sockets is
open at any given time instead of scaling the number of sockets linearly
with the number of ranks. The RingExchange algorithm has also been
replaced with an AllGather to simplify the code.

Change-Id: I1a830b23733003dfea21f0ff946ed687b9ce4673
Reviewed-on: https://git-master.nvidia.com/r/1813190
Tested-by: Wojtek Wasko <wwasko@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: David Addison <daddison@nvidia.com>
Reviewed-by: Wojtek Wasko <wwasko@nvidia.com>