Fix Bug 2127345 : Memory leak in ncclCommDestroy
The compare_and_swap code was not working, causing the barrier loop
to exit with a ncclInternalError, but since commFree return code
was not checked in ncclCommDestroy, we didn't see the failure.

This patch moves all intra-process barrier code into separate
functions and use those functions for the final barrier in commFree.

Also, this adds a check on commFree() return code inside
ncclCommDestroy.