Added extra debugging and sanity checks to RDMA code
We saw an issue in BUG 2132976 where the user was calling
ncclBcast() with different count arguments in each rank.
This caused NCCL to use the LL FIFOs in one process and the
non-LL FIFOs in the other. The resultant data overrun caused
an asynchronous IB error that was very hard to track down.
This change exchanges the buffer sizes between the ring
peers and so will hopefully catch this error earlier and in
a more comprehensible manner. However it will not catch all
cases where the user passes in differing arguments to the
collective calls.