Full rewrite of the search code
Instead of randomly searching the graph, precompute all GPU-GPU,
NIC-GPU and GPU-NIC shortest paths with a breadth-first search.
Then during the search hop from NIC to GPU to GPU to NIC following
the pre-computed paths. It is much faster and avoids situations
where we go through PCI between GPUs, while we have NVLink
connectivity which would forbid that.

The code it greatly simplified and permits much deeper searches
resulting in better results. Timeouts are still needed though,
in particular on P9 where there is no perfect solution so we try
to explore all paths.

But thanks to an elaborate sort of the next GPU to try, we should
always try to good solutions first, hence when we reach the timeout
we should already have the best solution saved.