Do not let netFactor decrease nChannels
Avoid the situation where the tree, with a netFactor of 2 finds a solution
with less channels than the rings. It would cause us to use less channels
since we take the minimum of the tree and ring solutions.

So, we first compute a solution without netfactor (which should have the
same number of channels than the rings) and then try to find better solutions
with the same number of channels but just higher speed.