Fix LL128 perf regression
This is a weird case where using sendStep[wid] actually helps
contrary to the LL case where is causes sendStep[i] to no longer
be in registers.

Register usage should be studied to better understand why we
suddenly drop from 100GB/s to 80GB/s with just that minor change,
using sendStep[wid] instead of sendConnHead, while both should be
in registers and equally fast.

Also in this patch :
 - only call checkAbort() when we don't have data to receive
 - only post tail when fifoPtr is no NULL.