Better search for non-GDR NICs
For NIC that don't support GDR, make the search go through
an intermediate CPU to try to find a better solution.

This is still imperfect as we don't reflect whether we use GDR
for sends in the case we do support GDR. To do so, we'd restore
the GPU<->NIC paths and force an intermediate step through the
closest CPU during the search.