Properly cleanup proxy ops in case of Group error
That caused hangs when we had to cancel group ops and used
the network transport. Also add a lastOpCount counter so that
we can rollback to the previous opCount when cancelling. That
could cause issues if cancel is inconsistent between ranks.
And use that lastOpCount to make sure proxies don't start
processing operations which might end up being cancelled.