Fix crash when GPUs on the same node have different GDR capabilities.
When mixing Geforce and Tesla, some GPUs would consider GDR is
available while others would not, resulting in different graphs
during search. This would cause discrepancies and cause crashes.

The fix discovers GDR capabilities for each GPU very early and
share it with others during the first bootstrap allgather. Then
we use that bitfield during the construction of paths so that
we reflect whether each GPU can use GDR or not.