Fix NUMA detection for NICs
To merge NICs with multiple ports into one, we trimmed the last 2
characters from the path (and modified the path permanently). This
is a problem since we won't then able to figure out the NIC NUMA
node with $path/numa_node. Systems where NICs were attached to a
PCI switch which had GPUs were unaffected, since the PCI->CPU path
was already created.

To fix this, only create a copy where the subdevice is forced to
zero and leave the path intact in the info struct.