Fix scatter/gather in "direct" mode.
To check whether we intended to use direct or not, we checked
whether directBuff was NULL or not. Problem is, with asymmetric
send/recv, directBuff (which is recvbuff) might be NULL even when
we intend to use direct as sender.

Add a DIRECT template param to make that intent explicit.

Also add direct mode support to tree/allreduce.