Channel connection rework
Duplicate buffers we send from buffers we receive to fix a random
crash when somehow send/recv overlap even though they're supposed
to be blocking.
Only transmit the number of connect info we need, not a range.
Allocate buffers in the stack to avoid mem leak, in particular in
case of error.