Fix latency regression.
Remove the inlining of the first operation completely. Pass the
device communicator as argument instead. To track the index, we
now keep it in the communicator and need to take the max of all
channels as some channels could skip indexes.

Regarding performance, it seems the compiler would load the
arguments into registers regardless of if/when we access them,
which was loosing 10us with 640 threads (effect on other operations
might be more limited because they only launch 64 threads on small
sizes).

Not inlining is causing an extra load to system memory, but
that load was already needed anyway as soon as we had more than one
channel, plus the load takes <2us which is still a net win.

Bug 200655790.