Fix crash when using GDR.
Now that steps could need a GDR flush or not (depending on whether
the size is 0 or not), we cannot increase transmitted immediately
if we don't use GDR. We need to have all steps go through the same
code, i.e. wait for the step with Flush to complete before we can
increase the step after that which doesn't need GDR. Marking the
request as NULL makes us skip ncclNetTest() completely and
immediately increase transmitted for 0-byte steps.