Fix crash when injecting a topology and using a sub part of it.
Since we fill data recursively from GPUs and NICs, CPUs or PCI nodes
which don't have GPUs or NICs we need in the communicator will have
some missing information, causing a crash.

The solution is to mark GPUs and NICs with keep="1" in the XML and
run a trimming phase where all branches which don't have keep="1"
will be removed.

Github issue #379.