Fix hang in corner cases of alltoallv
Bug 3197885

The chunkSize computation guarantees that ranks sending/receiving
use the same chunk size. It does not however guarantee they will
split the work on the same number of channels. If one operation is
coupled with another operation with a larger size, we could end up
using more channels on one side. Since the chunk size is the same,
only the first channels would have work to do; while others would
see the offset is higher than the size, hence the size would be
negative and those channels would do nothing.

The bug happens when we one channel which has exactly 0 bytes
to process. Because we implement a 0-byte send/recv as a
synchronization, one channel would wait for a synchronization with
a channel which is not running on the other side.

The solution is to set the size to -1 when the offset is non-zero
(we're not doing a true 0-byte send/recv) and the offset is larger
or equal to the size. That way, all extra channels will truly do
nothing.