Rework grouped proxy operations
Grouped proxy operation used to be different proxyArgs elements,
linked together in another dimension in the proxy scheduling chain.

Now they are part of the same proxyArgs, which has an array of sub
operations which are part of the same group.

This allows for better synchronization of those sub-operation, to
allow for aggregation or simply smoother operation.

It simplifies the scheduling of the proxy as well, although it makes
the Send/Recv proxy code mode complex (to allow for future
optimization).