Use NICs round-robin in 1PPN cases.
In a 2GPUs:2NICs per switch configuration, e.g. on DGX A100, when
creating communicators with one GPU per node, GPUs will only use
one NIC (as it's enough to get SOL BW) but they will always pick
the same NIC, so we'll end up with only half the NICs used.

This patch first selects only the NICs which are closest to GPUs,
then picks the first one to try based on the device number of the
first (and potentially only) GPU.

That way, a communicator with GPU 0 will try NIC 0 first, while the
one with GPU 1 will try NIC 1 first.

Bug 3018922.