Rework PXN path creation
Do not determine PXN path during the first path computation but
create them afterwards, selecting which GPU to go through based on
the ncclTopoGetLocalNet detection algorithm. This ensures we don't
use the same GPU for each pair of NICs on the same switch in DGX
A100.

This patch reverts some of the changes from the first PXN patch.