Rework GDRCopy API
Make the GDR Copy handle opaque and wrap the copy function.
Avoid manipulating gdr structs in the middle of the code.
Remove gdr include in devcomm.h to avoid recompiling all device
code when we change gdr.
Provide the cuda pointer as part of the GdrCudaCalloc call.
Allocate the GDR descriptor dynamically. Harmonize the code as all
GDR handles are now void*.