Fix performance regression due to the proxy list rework
With the new design, when running with -n 2000, we would append
a list of 2000 x nranks arguments to the posted list. Then during
ProxyAppendPosted, we would go through that list and add all
elements, and in particular starting with sends which would then
make our progress list much longer.

The solution is to only append args for the next operation based
on opCount.