Revert d519a5 as it breaks CUDA graph
Without the content of doneEvent being updated (say, by a non-graph
run), this is essentially calling a sync on an internal node of the
graph, so it may not be allowed.

After the event is updated by a real cudaEvenRecord call (say by a
non-graph run), one can sync on it again, and the already-captured
graph and launches of it would not be impacted.