Fix tree creation for ZionEX
With the balanced tree, the fallback on less good trees was broken,
as we would only try TREE from SPLIT_TREE but never from BALANCE_TREE.

Therefore on ZionEX we would never try to find a tree and given
BALANCED_TREE is not working well with the cubemesh topology, we
ended up with 7 channels which would probably limit rings to 7
channels down the line.

Adding ZionEX to graph unit test, with graph reference showing 8x24
on both tree and ring.