Add support for multi-operations
Add NCCL_TESTS_SPLIT_MASK env var to allow for perf tests of
communicator partitions.
Add group number in the initial list of GPUs.
Set MASK to a value in Hex so that the color will be rank & mask.
For example, 3 would split the ranks in 4 groups of N/4 GPUs, while
FFFFC would create N/4 groups of contiguous 4 GPUs.