Fix average option
The code was not compatible with the multi-thread mode.
Replace the MPI_Allreduce calls by an allreduce function performing
an inter-thread and inter-process allreduce.

Also fix weird usage of the deltaHost array, which seemed to have
all threads compute delta on the same part of the array, then perform
a max across different blocks (instead of different threads?).

Bug 200749134