Fix Memory leak of NVB connections
We need two IPC Ptrs: one for the memory we allocated (possibly on
an intermediate device) and one for the memory the other side
allocated (also possibly on another device).

We keep only one remoteId/memRank though as each side is responsible
for freeing the buffer it requested to be allocated by the
intermediate rank.