Fix hang on send/recv with 0-sized recv operations.
The ProxyAppend was supposed to be called on sends first, then recv.
To differentiate between sends and receives, the test was done on
sub->sendbytes being non-zero. That meant 0-bytes sends would not be
enqueued right away, and would wait until the second round.

In the case where we would have more operations than what the CUDA
kernel can handle in a single round, proxy operations would have
two opCount values (opCount and opCount+1), meaning when we
finally enqueue the zero-bytes sends, the proxyAppend pointer
would point to a second-round operation (opCount+1) which is
different from opCount, hence the zero-bytes send would be appended
as a third round of proxy ops instead of being part of the first.
This leads to a deadlock as those were not being progressed, and
corresponding receives would get stuck.

The fix consists in taking the recently introduced sub->sendChunkSize
field to differentiate between sends and receives. That field is non
zero regardless of the size of the send.

Bug 200778228