Fix hang with 16 GPUS per node.
We only process up to 8 send/recv per channel, but we need all ranks
from a given node to be ready to send a multi-recv. So, when nodes
have 16 GPUs, we will have the first 8/16 sends ready and the others
will not have started. Given the 8/16 are all different on each rank
because the schedule is rotating, there will be no destination out
of the 16 for which all 8 sends will be ready, causing a hang.

The fix is to leverage the fact we now use 2 channels for inter-node
communication and spread the sends on those 2 channels, guaranteeing
that all 16 ranks will progress.

If we were to support 32 GPUs systems, we'd need to increase the
number of channels used by inter-node communication to 4.