Fix topology detection on FB-V100
Add topology for FB-V100 and add it to graph_test.
Rework timeouts to lower them and permit more tries, so that on
FB-V100 we try 8x10.0 after finding 4x12.0. This is needed since
we added PXN which increased the number of tries we were doing and
was triggering global timeout on that platform.