Fix hang in ncclCommDestroy with multi-GPU + PXN
With PXN, the service proxy may wait until other ranks close
their connections. That means we cannot wait for the service proxy
thread to complete, in case ncclCommDestroy is called in a loop
in the case of single process/multi GPU.
Instead, defer waiting the service threads until the last GPU of
the process exits.