Add Param to disable direct access to other GPUs user buffers
A local GPU does not always have access to the next GPU user buffer
even when that next GPU is within the same process and P2P is enabled
between the two. The user buffer could be a host pointer, a remote
GPU pointer, or a buffer allocated with new CUDA APIs not mapping
buffers automatically on P2P-enabled GPUs.