Improve stability of PXN net device choice.
Choose which net device to use for PXN based on the local topology
instead of the remote net device number. Different node can have a
different numbering if there are extra NICs up (e.g. attached
to the CPU, and some are up or down depending on the node).

Use the net device close to the local GPU with the same NVML number
as the remote rank.

Bug 3613299