Fix scatter/gather/alltoall performance regression
The channel computing was changed to spread the base and add inc
instead of spreading the inc and adding the base.

Problem is the comm->p2pChannels bit-mirrored array was only set for
the first p2pnChannelsPerPeer, which means on multi-node, only the
first two entries would be non-zero, all the rest being zero.

Bug 3646869