Fix alltoall performance on medium sizes on multiple nodes.
Since 2.12, we no longer use more than 2 channels per node. Hence,
we need to start using 2 channels as soon as possible to get peak
BW.
Recent alltoall tuning for intra-node made the max stepSize even
bigger, making the 2-node perf even worse.

Separate the per-channel stepSize tuning between single node and
multi-node; keep the same tuning for intra-node, and limit the
stepSize per channel to the real network chunk size (128K).

Performance is still not as good as 2.11 on two nodes, because we
only use 2 channels instead of 16, but it is much closer and the
advantage will invert as we scale.