Restore caching of gdrSupport
This is mostly to accelerate CI/CD on DGX-2, because we are still
running with CUDA 11.2/r460 and calls to RM slow down horribly as
we map more memory. This causes ncclCommSplit tests to take ~170s
each while they would only take 11s if alltoall had not been run
before.