Merge of https://gitlab-master.nvidia.com/nccl/nccl/-/merge_requests/158
* Bug fix: We now support interleaving graph captures on a communicator.
Previously for a fixed comm, capturing a NCCL op against graph A, and then
an op on graph B, and then another op against graph A would probably deadlock.
Now this works.

* New apitest: ncclCudaGraph_test/0.many_graph_many_stream_interleaved
This stresses the above bug fix.

* Perf regression fix: Graph capturing no longer uses empty graph nodes to
manage dependencies during ncclStrongStreamWaitStream(). This had a big impact
on perf, notably got DeepCAM back to its NCCL 2.12 performance. @David Fontaine
has confirmed that empty nodes are a known slowdown which has been fixed.

Summary of empty node change:
The graph NCCL was building contained empty graph nodes (cudaGraphAddEmptyNode)
to make handling dependencies easier. Internally we build a linear chain of
graph nodes (to emulate a stream). The old code kept track of the last added
node so it could make it the sole dependency of the next appended node. Sometimes
we have to make one chain sync with another (similar to syncing one stream to
another using cudaStreamWaitEvent). Since we only kept track of a single node,
we brought in the other chain by adding an empty node that had 2 dependencies, one
on this chain, one on the other. The change that restored perf was to keep track
of a list of nodes at the tip of the chain. Usually that list has just one node
like before. But when we sync against another chain, we concatenate this chain's
list with the other chain's list so that the next node added depends on their
union = no more empty nodes.

Fixes https://nvbugs/3726645
BUG 3726645