Fix crash in CollnetChain
CollnetChain was using the tree graph instead of the collnet graph, causing crashes
as the connection to SHARP was not initialized on some ranks. Heads of collnetDirect
are taken from the heads of the chains computed in the collnet graph, so collnetChain
not using the right graph was causing inconsistencies in rare cases where trees and
collnetChain differ.