Delete unnecessary last rank free logic in commFree()
We rely on last rank which calls ncclCommDestroy() to join all proxy threads in
the process; however, this "last rank" logic is implemented in commReclaim() now,
and it becomes redundant in commFree(). In addition, accessing intracomm0 in
commFree() is dangerous since intracomm0 might be freed already. This patch solves
this issue.