BUG 3726645
Improve number of task dependencies emitted in CUDA graphs. This greatly
improves performance of large CUDA graphs.

The issue stems from our use of an internal per comm stream for serializing
kernels and how dependencies between that stream and the user stream (which
gets the actual kernel launch) compound quadratically in the CUDA graph.

The dependencies we want to see is for kernels K we should have:

  K[0] -> K[1],
  K[1] -> K[2],
  ...,
  K[n-1] -> K[n]

But instead we were getting:

  K[0] -> K[1]
  K[0], K[1] -> K[2]
  K[0], ..., K[n-1] -> K[n]

Which is the transitive closure of the former, quadratically larger, but
technically not any worse in that there aren't any false dependencies and the
graph could *conceivably* schedule the same way. But it doesn't. Removing the
extraneous dependencies and presenting CUDA with the more compact
representation retrieves the performance.

The culpable code was in the interplay between the internal stream and the user
(a.k.a launch) stream. Before launching, we sync the user stream with the
internal stream, and then after launch sync the internal stream back against
the user stream. In graph mode, each sync is a set-union: the set of nodes
pointed to by the target stream has the set of nodes from the source stream
unioned into it. Since the internal stream was never getting any tasks launched
on it, nothing was causing that set to collapse back to a single node, so it
just grew without bound. The fixing change was to add a boolean to the sync
operation indicating that we know the source stream already subsumes all the
work of the target stream, allowing the sync to just replace the target node
set with the source node set. See changes to `ncclStrongStreamWaitStram`.