Move intraComm0 compute before proxy create
Moving intraComm0 compute before proxy create makes it convenient to
determine if we need to perform last-destroy logic when abort comms.
If intraComm0 is NULL, it means proxy thread is not created and no
connection is built up, so there is no need to let last thread to
free everything.