Fully asynchronous and complete init-time connection establishment
This MR contains a rework of the transport connection orchestration, and as a side-effect, asynchronous proxy op execution code.

The proxy service thread now contains a map of outstanding async proxy Ops per-localPeer (not send/recvs, but connect and setup calls.)  It iterates and progresses each request in a nested for loop.  This is opposed to the current design which has a 1-dimensional array of operations, each localPeer limited to a single slot.

Client threads interacting with the proxy have an option of calling async or blocking (before, all ops were blocking on the client side.)  An op ID is allocated by the client and passed to the proxy for the request. This op ID is returned to the client as part of the response.

The purpose of this work is to complete connections at initialization time, eliminating the need for runtime flow control (see ncclSendCheck().)  This aligns the design with the long term goal of GPU-initiated communication in which there will be no CPU-side proxy code to check if a sendComm is ready to send or not.

A desired side-effect of this change is faster connection establishment in general - all of the send and recv connections from every rank will be queued and serviced in parallel as opposed to serializing per-proxyLocalPeer. TBD on benchmarking of connection establishment speedup. This will also cut down on benchmarking noise since send queues are ready-to-send from the very first send (not as big of a deal but doesn't hurt.)

Asynchronous transport connection establishment. Unbounded asynchronous proxy operations. Send connections are now ready-to-send before leaving send connect() calls

See merge request nccl/nccl!183