Use all NICs for alltoall when we have more NICs than GPUs
This includes both the PXN and non-PXN case.
Remove the "preferred netdev advertising" for each rank, given we
were no longer using it (only local preferred NICs).
Detect which GPUs are local to each NIC and perform a round robin
on those local GPUs and local NICs to assign a NIC to a GPU/Channel
combination, and reversely, also assign a GPU to each NIC.