Fix NVB disabled case.
Since we added the patch to use the network instead of SHM when
performance is better through the network, NVB disabled would get
very bad performance with Tree on DGX-1.

That's because SHM would return "0" when asked whether non-NVLink
connected GPUs can connect through SHM, and therefore we'd mark
path.count = 0 so that the GPUs are later disconnected. Problem is
GPUs would not be disconnected because they're still connected
(indirectly) through NVLink, so the paths between GPUs not having
a direct NVLink would keep path.count = 0, bw=6GB/s and type=SYS.

And the search would then not even check the path type and bw;
if count is 0, it would consider the link is always good to use.
We'd end up with Tree going from GPU 3 to GPU 4 on Tree channel 0
which would go through the network and not even use the right NIC
because Tree channel 0 is using NIC 0.

The fix is to not set path.count = 0 but instead mark path.type=NET.
Then, when disconnecting GPUs, check that type is NET instead of
count = 0. Then during the search, PATH_NET would always fail to
be used because at most we try to create paths with PATH_SYS.