Support multi-thread Fault Tolerance
If NCCL communicator and op are created and launched by one thread
but aborted by another thread, we might fail to stop nonblocking
op because the group op is tracked by thread-local variables. This
commit makes these variables linked to communicators so that we
don't need to worry about which thread creates and aborts the
NCCL communicators and ops.