Fix of performance regression for ncclGroup's with 2 or more non-allreduce collectives.
Accounting of bytes-of-work within a ncclGroup was wrong for non-allreduce collectives since
it was using "ncclInfo::count" which is the per-rank buffer size for allgather and reducescatter.
Now we use "ncclInfo::nBytes" which is all-rank buffer size which better reflects the amount of
data being processed.

See merge request nccl/nccl!246

BUG 4138474