Fix search for Viking/SHARP/4PPN
H800 fixes made the search duplicate the channels when nChannels
was 4 or less.

That broke the search on 4PPN which fell in that case for the ring
search, but due to an old feature (no longer valid) limiting the
usage of each NIC to 1 when SHARP is enab;ed, the Tree search could
not find a similar solution with 8 channels x 24 (instead of 4 x 48).

Remove old net.maxChannels parameter since we now share resources.
Fix collnet/Direct when we have duplicated channels in the search
graph.
Add Viking-4GPUs-SHARP topology

Bug 4146716