Ensure consistent topology regardless of the GPU list.
When creating 1PPN communicators on a DGX/HGX system, we might end up
with inconsistent topology. This is because we first add GPUs, which
will also add the PCI switch topology tree, then we add NICs, and
then we flatten BCM switches.

So when we add only one GPU, NIC 0 and 1 may end up being reversed
if NIC 1 attaches to the same sub-switch as the GPU, after the BCM
flattening.

This patch orders the PCI tree based on PCI IDs to avoid the risk of
reordering.

Github Issue #820
Bug 4201618