Fix aggregation performance with collnet
Sort proxy ops by opCount before submitting them to the net proxy.
Collnet proxyArgs aggregation relies on operations with the same
opCount across different channels to be recognized as part of the
same network operation.

Tune the number of channels to use all channels for each operation
rather than splitting them on different channels, as it would not
allow for aggregation.