Fix cudaMemcpyAsync bug
See Github issue https://github.com/NVIDIA/nccl/issues/957

We are trying to use the copy result of first cudaMemcpyAsync in the
second cudaMemcpyAsync without sync in between. This patch fixes it
by allocating a CPU side array to cache device side addr so that we
can avoid this consecutive cuda mem copy.