Fix alltoallv crash in self-sendrecv.
Probably also avoid hangs in some situations.

The fuseOk variable was supposed to avoid fusing operations from
different nodes, and make sure we use a new work element when we
communicate with ourselves, because the send and recv operations
for self are supposed to be next to each other (typically in
position 0 and 1 of the ncclWorkElemP2p).

But fuseOk was only set right for the first chunk, meaning if we
were to use multiple channels (multiple chunks), then we could end
up with unwanted fusing.