Modify -R option to separate Send and Recv local registration.
Change the 0/1 option to s/r/a for send/recv/all buffers. This is
important to test p2p registration properly and should also be
useful for NVLS to test corner cases.

If we don't see s/r/a as -R arg, still call strtol to be backward
compatible with 0/1 arguments.