Squash commit for IB Port Fusion
This MR implements physical NIC merging in NCCL

The purpose of this is to aggregate the bandwidth of two NICs under one logical NCCL network transport connection. This makes it easier for NCCL to make use of all "NICs" on dual-port IB systems, which expose one physical 200Gbps NIC as two 100 Gbps NICs to the application.

The IB transport was mostly rewritten for this, the reason being we must decompose the concept of one IB device per-NCCL netComm, which was baked in all over. There are now one or two IB devices, connected to one or two remote IB devices, with QPs spread among both.

Merging in init(), getProperties()

See merge request nccl/nccl!337