Revert singleChannelDBTree feature
The branch was causing a huge memory increase.

Intra-node, each channel would allocate two buffers for the tree
algorithm, one for each connIndexes.

Inter-node, the shared buffers would be used by both coll and p2p,
and had to be sized for ring/tree, but multiplied by the number of
channels and the number of localRanks for PXN.

That feature also enabled the increase of default channels to 24,
which was also increasing the memory usage by 1.5x.