Fix hang with multi-QP + registered receive.
In the case of multi-QP, the size returned by the plugin was
multiplied by nqps. This wasn't noticed until 2.19 because we only
use the size to know whether it was zero or not (and call flush).

In 2.20 we rely on it in the case of registered buffers to know
what we received and potentially add more receives.