Fix collnet operation
The new intranode allgather algorithm only worked in cases where
all ranks were synchrnous, and do not have outstanding messages.

In the case of collnet, it seems we send a handle to ourselves,
then perform an allgather, then receive that handle. The outstanding
send to self will break if the allgather is not using the standard
bootstrapSend/bootstrapRecv functions, as they use the same listen
socket, and therefore we need to put the send to self in the
unexpected queue.