Fix packet reordering issue in bootstrap.
bootstrapSend/bootstrapRecv open a new connection and close it for
every send/recv. Not being part of the same socket, we have therefore
no guarantee that two messages sent between a pair of ranks will
arrive in order, and it is frequent on AWS to see messages passing
others.

For that reason, we added tags, so that communication between a given
pair of ranks would match in order. The barrier added after the
connectInfo exchange was using the exact same tags as the connectInfo
phase, and was therefore prone to message reordering causing a
mismatch.

Bug 4497047