Fix device unpack peer mask
To know for which peer we need device unpack, we compute a mask,
where each bit set corresponds to a peer index.
That mask is computed using __ballot_sync, hence we need threads
to set their value to "1" only if their warp id is equal to a
peer index which uses unpack.
Fortunately, threads with a recvWait role fit that property. They
start from tid 0 (hence warp id 0), and they load the connection
for the peer equal to their thread ID (=warp id) so they can know
whether their peer needs unpack or not.

The code is setting the NetDeviceUnpack flag for both roleWaitRecv
and rolePostRecv. Therefore, if the rolePostRecv thread
ended up in the first warp, we would have a mask where the bit
corresponding to the rolePostRecv thread would be set, causing
crashes.

To solve the problem, we simply only set the flag in the ballot_sync
if our role is roleWaitRecv, since this is the only role where our
warp ID makes sense as a peer index.