Fix NVLSTree max chunksize
The computation of maxChunkSize was broken if the NVLSTree max
chunksize was larger than the NVLS chunksize, and we would end up
with data corruption.

To fix the issue, start with the NVLS chunksize and then reduce it
if we compute a lower max chunksize (which can be overriden by the
env var).