Prevent infinite recursion in findLocalCpu for c2c
findLocalCpu has a tendency to go into an infinite recusion for a
certain Node Topology.

A sample node topology which could trigger such a case is :
<system ...>
  <cpu ...>
    <pci ...>
      <pci ...>
	<nic>
	  .
	</nic>
      </pci>
      <pci ...>
	<nic>
	 .
	</nic>
      </pci>
      <pci ...>
	<gpu ...>
	  <nvlink .../>
	  <c2c .../>
	</gpu>
      </pci>
    </pci>
.

During the second pass during topology detection phase when the
xmlToSystem code encounters a c2c link it backtracks and tries to find
the local CPU from the GPU node.

While it tries to backtrack it traverses the links in the graph where
it encounters the PCI Switch.  When exceution comes to the PCI switch
it traverses its links and lands back again on the originating GPU
thus creating an infinite recursion.