Squash commit for Disable port fusion in heterogenous systems
Reset and rebuild IB device list with port fusion disabled if a mix of multiport and single port NICs are detected. The current port fusion code can't work with a single-port on one side and dual-port on the other for now.

BUG 4607459

Reset and rebuild list with port fusion disabled if a mix of multiport and single port NICs are detected

See merge request nccl/nccl!442