tuner: fix bug in tuner_v2_as_v3 compatibility layer
In the tuner plugin v2 interface the plugin was allowed to return
any algo/proto combination and NCCL will accept it. With the v3
interface this is no longer the case as the returned algo/proto are used
to index a value in the cost table. Therefore, now we need to validate
that the algo/proto fit into the range of valid combinations and they do
not overwrite a combination that NCCL considers invalid.

Also fix some typos.

NVBug https://nvbugspro.nvidia.com/bug/4661598