bootstrap: performance optimization and scalability API
Introduce the following optimizations:
- merge of bootstrapAllgather calls to reduce latency
- enable the root to send information upon reception and not wait for all the information to have been received
- enable the use of ncclNet to perform the global allgather
- add a new API for scalable init

1. merge bootstrapAllgather calls
Merging calls to bootstrapAllGather improves performance by reducing the number of global rings that have to be made

2. early return
The root is currently waiting to receive all the information from all ranks before sending the next address to each of them.
This creates a sync event and most ranks are waiting for the root.

This optimization enables the root to send the correct information to the rank upon reception
- if the information just received can be sent to a rank, send it
- if the information needed by the rank just received, send it
In both cases, we will not store the received information.

Once all the information has been received, we proceed to a final sweep and send all the information that couldn't be sent early

3. use of NCCL_NET for out-of-band communications
The usageof NCCL_NET for OOB comms can be ENABLED with
NCCL_OOB_NET_ENABLE.
NCCL will then use the ncclNet interface to complete the global
Allgather ring

4. scalable API: many roots: add many roots implementation for bootstrapInit

To improve the scalability issue of the handshake (single root, many ranks), we add the ability to use more than a single root for the initial handshake.

- the number of roots must be >=1 and <= number of ranks
- each rank will send its information to the associated root. The association for `P` processes and `R` roots is done the following way: the first `P%R` roots are associated to `P/R + 1` ranks. The remaining roots are associated to `P/R` ranks.
- the first rank of each root needs to send its connection information to the associated root and the previous root. The associated root needs to reply with the handle of its neighbor and therefore requires the connection info. The previous root needs to send its connection information to the last rank associated to the root.
- once the next address is received, we build the ring and perform the allgather to obtain all the connection information of everybody

Implementation details:

- the user-provided ncclId need to be copied locally in case async communicator creation is used. Because we copy to an allocated memory, we can remove the union in the job structure.
The alignement of calloc will be compatible with both ncclUniqueId and ncclBootstrapHandle, hence not causing issues when casting one into the other

- the commHash needs to be obtained differently in init.cc if we are creating a new comm or splitting one (cfr below)

- `NCCL_UNIQUE_ID_BYTES` is larger than `sizeof(ncclBootstrapHandles)`. Therefore care must be taken when casting the `ncclUniqueId*` to a `ncclBootstrapHandle*` as the displacements are obtained based of the types that have different sizes.
To avoid this issue, `bootstrapInit` now takes a void * which allows to obtain the right array access using the macro BOOTSTRAP_HANDLE

- `BootstrapSplit` doesn't have access to the a valid `ncclBootstrapHandle` anymore which can be used to obtain the magic or the hash.
The `BootstrapSplit` function now takes the magic to be associated with the communicator, as it's the only information that is needed.

- the magic logic is moved to `init.cc` and is as follows: previously the commId (whose first 64 bits are the magic) was created by taking the first 16 bytes from the parent's commHash, and adding the color.
Now, the magic used for the bootstrap is the same as parent's to avoid any issue when reusing sockets. We use the commHash value as a magic for simplicity.

- the commHash of a comm in the case of splitting is now computed using the same mechanism, using a temporary location.

- we remove a few conditional and use a function name array instead to simplify the code

many roots test/api: add API test for the API extension

many roots test/mpi: modify the mpi_perf test to use scalable comm init

- MPI_PERF_N_ROOTS determine how many roots are used to init the comm
- the ncclUniqueIds are distributed using MPI_Allgatherv to all the processes

Further, we also have other smaller improvements:

- bootstrap: adapt the throttle on the client side

To avoid overflowing the root with msgs we adapt the message rate when
there is a large number of ranks. The targeted message rate is 7k
msgs/sec which corresponds to typical observed value.

The delay is obtained using the local_id and the root associated to the
rank id. We neglect the case of sending a message to the previous root
as it's only a single message and it should be negligible compared to
the bulk of messages associated to the other processes.

The threshold above which delay is added can be changed with
NCCL_UID_STAGGER_THRESHOLD env variable. The targeted message rate can
be set by NCCL_UID_STAGGER_RATE.

- Further we also add an additional option to use the scalable API in
the per/test

Use `-I` or `--init_ids` to use the new scalable API ncclCommInitRankScalable with more than one ncclUniqueId when creating the communicators in nccl perf tests

- if unspecified, the old way is used (`ncclCommInitRankConfig`).
- if a value of 0 is given, we use 1 ncclUniqueId per 128 NCCL ranks
- if the value given is larger than the number of ranks, we use the number of ranks

Then, MPI_Allgatherv is used to broadcast the different Ids accross processes.